Renamed codeable_concept -> CodeableConceptUtils and moved into model directory.
